Forex projects added.

+++ FinancialData.Shared/IBar.cs
@@ -21,15 +21,15 @@ using System.Text;
 using System.Threading.Tasks;
 
 namespace FinancialData.Shared {
-	public interface IBar : ISample {
-		Asset Asset { get; }
-		IDataSource Source { get; }
-		DateTime DateTime { get; }
-		decimal Open { get; }
-		decimal High { get; }
-		decimal Low { get; }
-		decimal Close { get; }
-		int Volume { get; }
-		decimal AdjClose { get; }
-	}
+  public interface IBar : ISample {
+    Asset Asset { get; }
+    IDataSource Source { get; }
+    DateTime DateTime { get; }
+    decimal Open { get; }
+    decimal High { get; }
+    decimal Low { get; }
+    decimal Close { get; }
+    long Volume { get; }
+    decimal AdjClose { get; }
+  }
 }
